Make Device loadable from custom file
The Device was being loaded from a YAML description in the previous
version. Now the class is specified in a YAML file but the implementation
of the Device can be in any module.

Subclassing a Device will allow us to override its behaviour and
implement much more complex instrument devices.